backend/audiobackend/requirements.txt
django
djangorestframework
django-cors-headers
librosa
numpy
scipy
matplotlib
SpeechRecognition